feat: divide Pryzm Protocol into two protocols
